+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Warning when deleting datasets which are out of sync with their push remotes
+
 ## [0.32.0] - 2024-11-28
 ### Added
 - Query page: added default query after searching for first dataset